Consolidate helper methods that go ThroughTheCore and ThroughTheWeb into modules
