+++ test/core/advice_test.exs
@@ -1,6 +1,6 @@
 defmodule Advisor.Core.AdviceTest do
   use Advisor.DataCase
-  alias Advisor.Core.{Answers, Advice}
+  alias Advisor.Core.Advice
   alias Advisor.Core.Questionnaire.Creator
   alias Advisor.Web.QuestionnaireProposal, as: Proposal
 
@@ -9,14 +9,14 @@ defmodule Advisor.Core.AdviceTest do
   test "can figure out if an advice has been answered fully" do
     christoph = advice_with("Christoph Gockel", @questions)
 
-    answer!(christoph, with: %{"1" => "foo", "2" => "bar", "3" => "batz"})
+    ThroughTheCore.answer!(christoph, with: %{"1" => "foo", "2" => "bar", "3" => "batz"})
     assert Advice.completed?(christoph, length(@questions))
   end
 
   test "cann tell if advice has not been answered fully" do
     christoph = advice_with("Christoph Gockel", @questions)
 
-    answer!(christoph, with: %{"1" => "foo"})
+    ThroughTheCore.answer!(christoph, with: %{"1" => "foo"})
     refute Advice.completed?(christoph, length(@questions))
   end
 
@@ -26,9 +26,5 @@ defmodule Advisor.Core.AdviceTest do
                                                        group_lead: "Jim Suchy",
                                                        questions: questions) |> Creator.create
     individual
-  end
-
-  def answer!(%{id: id}, [with: data]) do
-    Answers.store(Map.put(data, "id", id))
   end
 end
