+++ test/core/questionnaire/deleter_test.exs
@@ -14,20 +14,12 @@ defmodule Advisor.Core.Questionnaire.DeleterTest do
 
     {:ok, %{questionnaire: id, advisories: advisories}} = Creator.create(proposal)
 
-    answer!(advisories, with: %{"1" => "Foo", "2" => "Bar"})
+    ThroughTheCore.answer!(advisories, with: %{"1" => "Foo", "2" => "Bar"})
 
     Deleter.delete(id)
 
     assert [] == Answers.find(advisories)
     assert [] == Advice.find(advisories)
     refute Questionnaire.find(id)
-  end
-
-  def answer!(advisories, data) when is_list(advisories) do
-    advisories
-    |> Enum.each(fn(advisory) -> answer!(advisory, data) end)
-  end
-  def answer!(%{id: id}, [with: data]) do
-    Answers.store(Map.put(data, "id", id))
   end
 end
