+++ test/core/summary_test.exs
@@ -3,7 +3,7 @@ defmodule Advisor.Core.SummaryTest do
 
   alias Advisor.Web.QuestionnaireProposal, as: Proposal
   alias Advisor.Core.Questionnaire.Creator
-  alias Advisor.Core.{Answers, Summary}
+  alias Advisor.Core.Summary
 
   setup do
     proposal = Proposal.build(for: "Rabea Gleissner",
@@ -14,19 +14,11 @@ defmodule Advisor.Core.SummaryTest do
     [proposal: proposal]
   end
 
-  def answer!(advisories, data) when is_list(advisories) do
-    advisories
-    |> Enum.each(fn(advisory) -> answer!(advisory, data) end)
-  end
-  def answer!(%{id: id}, [with: data]) do
-    Answers.store(Map.put(data, "id", id))
-  end
-
   test "presents tabular data for a given questionnaire", %{proposal: proposal} do
     {:ok, created} = Creator.create(proposal)
     %{advisories: advisories, questionnaire: id} = created
 
-    answer!(advisories, with: %{"1" => "Foo", "2" => "Bar"})
+    ThroughTheCore.answer!(advisories, with: %{"1" => "Foo", "2" => "Bar"})
 
     [_headers, first, _second] = Summary.for_download(id)
     assert length(first) == 5
