+++ test/web/controllers/progress_page_test.exs
@@ -22,7 +22,7 @@ defmodule Advisor.Web.ProgressPageTest do
                             |> Links.generate
 
     conn
-    |> login_as("Felipe Sere")
+    |> ThroughTheWeb.login_as("Felipe Sere")
     |> get(progress_page)
     |> html_response(200)
     |> has_requester("Rabea Gleissner")
@@ -37,11 +37,11 @@ defmodule Advisor.Web.ProgressPageTest do
                                            |> Links.generate
 
     conn
-    |> login_as("Chris Jordan")
+    |> ThroughTheWeb.login_as("Chris Jordan")
     |> post(link, ["1": "someting"])
 
     conn
-    |> login_as("Felipe Sere")
+    |> ThroughTheWeb.login_as("Felipe Sere")
     |> get(progress_page)
     |> html_response(200)
     |> has_completed_advice()
@@ -56,15 +56,15 @@ defmodule Advisor.Web.ProgressPageTest do
     answers = ["1": "something", "2": "else"]
 
     conn
-    |> login_as("Chris Jordan")
+    |> ThroughTheWeb.login_as("Chris Jordan")
     |> post(cj, answers)
 
     conn
-    |> login_as("Priya Patil")
+    |> ThroughTheWeb.login_as("Priya Patil")
     |> post(priya, answers)
 
     conn
-    |> login_as("Felipe Sere")
+    |> ThroughTheWeb.login_as("Felipe Sere")
     |> get(progress_page)
     |> html_response(200)
     |> has_continue_button_with("We are good to go")
